Move fixtures.py to fetch.py
